hoja_de_vida/requirements.txt
# Django and core dependencies
Django>=4.2,<5.1

# PDF generation
WeasyPrint>=63,<65

# Image processing
Pillow>=10.0,<11.0

# Azure Blob Storage
azure-storage-blob==12.19.0
azure-identity==1.15.0

# PDF manipulation (for merging certificates)
pypdf==4.0.1

# Static files and deployment
whitenoise==6.6.0
gunicorn==21.2.0

# Additional utilities
requests==2.31.0
python-decouple==3.8
